requirements.txt
Flask==3.0.3
python-telegram-bot==13.15
python-dotenv==1.0.1
gunicorn==23.0.0
requests==2.32.3
uvicorn==0.30.6
aiohttp==3.9.5